Update cen4072.c
This are some of the subtle bugs that exist, like I said it is about submitting something in. 
line 49 should be "%d" instead of "%c"
line 58 causes a fatal error if user picks 3
line 96 should be "%d" instead of "%f"
line 99 left "fwrite" function where remove should be. 
line 100 for loop is unreachable
line 101 should be name = not name ==